Updated example cases, fixed bugs in the Plumed plugin
Features
--------
Input file keywords have been added for setting velocities through the
command line interface.

There are now two up-to-date example cases: one showing how to perform a basic QM/MM
simulation using both the Python API and the CLI (case_0), and one showing how to
perform a basic simulation using a restraining potential with the Plumed plugin
(case 1).

Bug Fixes
---------
Two minor bugs in the Plumed plugin were corrected.  The default arguments for
`return_forces` and `return_components` were set to true, since the plugin requires both.
The `biased_energy` is returned as a `numpy.ndarray` with a single element, so the value
of that element is now passed into the energy and components attributes of the `Results`
object, as opposed to the array object itself.