Streamlined input file loading, multiple bug fixes
Features
--------
The `Atom` object is now defined in the `common` directory, since the object is
a more general data container than the `System` object.

The `FileManager` class has been removed in order to follow a Singleton design
pattern--all file management is now performed by functions within the
`FileManager` module in the `common` directory.

System loading now only requires PDB files.  Consequently, XML reading
utilities have been removed.  The XML utilities existed to extract partial
charge and mass data from FF XML files, but masses are now instantiated from
standard molecular weights defined in the `common/units` module.  Partial
charges are initially set to zero.  Masses and charges in the system object
are now overwritten in the `MMInterface` factory method using data from the
forcefield file.  The `System.load` method now takes an arbitrary number of PDB
files, so long as the lattice parameters are equivalent between the PDB files.

Arbitrary triclinic boxes are now supported.  Accodingly, a utility to compute
lattice vectors, `compute_lattice_vectors`, from lattice parameters has been
added.  The `Wrap` and `Center` plugins have been updated to support arbitrary
triclinic boxes.  Additionally, the number of PME grid points to use across
each lattice vector may now be specified as a tuple of three integers when
instantiating the `MMHamiltonian` class.

Support has been added to `mm_hamiltonian` and the `openmm_factory` modules
for: AMBER prmtop/parm7 files; CHARMM psf, par/prm, rst/top, str, and inp
files; and GROMACS top files.  These files are read in using the ParmEd package
as an optional dependency.  Accordingly, the `forcefield_file` and
`topology_file` arguments of the `MMHamiltonian` class have been replaced with
a single `forcefield` argument, which may be a string or list of strings,
depending on the number of forcefield/topology files needed to describe the
system.

Utilities have been re-organized into different modules within the
`common/utils` folder.  This was done to prevent a bloated `utils` module.

Support has been added for mixed basis set definition.  This is accomplished
through a call to the Psi4 `basis_helper` function in the `psi4_utils` module.
Use of the mixed basis set functionality is achieved by assigning a multi-line
string to the `basis_set` argument when instantiating the `QMHamiltonian`
class.  Each line of the multi-line string must be of the form `assign
*element symbol* *basis set name*`, omitting the asterisks.

A selection query string may now be passed to the `System` object when indexing
with brackets.

Bug Fixes
---------

A minor bug in the `Stationary` plugin was fixed where a frozenset would be
erroneously used to slice an array.

A check has been put in place to ensure that `NaN` or `Inf` values will no
longer be written to output PDB or DCD files.

OpenMM NonbondedForce exceptions are now updated when charges are updated.
Most of these exceptions either set the charge product to zero or to some
scaled value of the true charged product (as in the case of 1-4 interactions in
many forcefields), so the appropriate scaling is now applied when charges are
updated in the `System` object.

Previously, coordinates were output to DCD files which centered on Subsystem I,
but now un-translated coordinates are output to DCD files.  This bug fix does
not yet apply to coordinates output to PDB files.

When a mass is set to zero by a forcefield, OpenMM interprets this atom as
being stationary.  Previously, these zeroed masses would cause the PyDFT-QMMM
integrator to crash, since it requires non-zero masses.  This bug has been
corrected by allowing the `Simulation` object to generate a query for all atoms
with mass zero and instantiating and applying a `Stationary` plugin object to
the integrator.  The zero masses are then overwritten with standard molecular
weights defined in the `common/units` module.