Adjusted partition criteria and non-periodic boxes
Features
--------

`ParitionPlugin` objects are now able to include atoms within the same
residue(s) as the atoms comprising subsystem I in subsystem II.  This is
done in anticipation of forthcoming implementations of link-atom and
local SCF approaches for handling subsystem boundaries across covalent
bonds.

Bug Fixes
---------

A bug which required the `System` object to have periodic lattice
vectors when the nonbonded method was set to `NoCutoff` or
`CutoffNonPeriodic` has been fixed.