Improved SETTLE and constrained optimization
Features
--------

An inefficiency in the `SETTLE` object causing slow generation of water
residue data has been eliminated.

The `Optimization` object can now perform constrained optimizations.
To perform a constrained optimization, the user must write the
GeomeTRIC-style constraints to a text file in the working directory,
and then provide the name of the text file to the `constraints`
keyword argument when instantiating an `Optimization` object.  The
`Optimization` object also now uses the internal conversion factors
of PyDFT-QMMM.

Support for specifying a `CustomBondForce` in OpenMM forcefields has
been added to the `openmm_factory.py` and `openmm_utils.py` modules in
the `interfaces/openmm` sub-package.

OpenMM exclusions for QM atoms in the `openmm_utils.py` module of the
`interfaces/openmm` sub-package have been generalized such that
exclusions are applied to a term if its atoms form a subset of the QM
atoms.  Previously, exclusions were applied if the term atoms and QM
atoms were not disjoint.

The `Molecule` object generated by the `Psi4Interface` object now
includes atom labels for each atom, corresponding to the atom’s index
in the `System` object.  This is useful for applying different basis
sets to atoms of the same element.  C1 symmetry is enforced for the
`Molecule` object.

The `residue_map` property of the `System` object now returns an
immutable `MappingProxyType`.

Bug Fixes
---------

A bug which would assign an empty box vector to an OpenMM `Topology`
object has been fixed.

Non-periodic systems will no longer be written with a `CRYST1` line
when using the `write_to_pdb` function.